requirements.txt
qrcode
opencv-python
pillow
pyzbar
